feat: Moved DateContainer out of components
